feat: Add TypeScript to Lua transpilation setup
This commit introduces a setup for transpiling TypeScript to Lua. It
includes a new .gitignore file to ignore generated Lua files, a
deno.json file for defining tasks such as build and clean, and a
deno.lock file for locking down the versions of dependencies.

A new TypeScript file, main.ts, is added which contains a function to
parse a file into a certain structure. A tsconfig.json file is also
added to configure TypeScript to Lua transpilation options.